third_party/blink/web_tests/fast/events/mousemove-from-iframe-to-top-element-expected.txt

TOP

Tests that hover is lost from iframe when the mouse is moved from iframe to the element that stacked up on iframe

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS successfullyParsed is true

TEST COMPLETE

PASS hoverBgColor is "rgb(255, 0, 0)"


PASS events[0] is "mouseenter"
PASS targetIds[0] is "target"


PASS events[1] is "mousemove"
PASS targetIds[1] is "target"


PASS bgColor is "rgb(255, 255, 0)"


FAIL events[2] should be mouseleave. Was mouseout.
PASS targetIds[2] is "target"


FAIL events[3] should be mouseout. Was mouseleave.
PASS targetIds[3] is "target"
